Mejoran los giros
Incluido un PID preeliminar. Seguir afinando. Pero va resolviendo. Faltan bifurcaciones de ir adelante o girar a un lado, y la de poder ir hacia adelante o ambos costados.

+++ Core/Inc/PID.h
@@ -21,10 +21,10 @@ typedef struct {
 
     int32_t outputMin;
     int32_t outputMax;
-    int8_t base;
+    uint8_t base;
 } _sPID;
 
-void PID_Init(_sPID *pid, int32_t Kp, int32_t Ki, int32_t Kd, int32_t min_max, int8_t base);
+void PID_Init(_sPID *pid, uint16_t Kp, uint16_t Ki, uint16_t Kd, int32_t min_max, uint8_t base);
 int8_t PID_Compute(_sPID *pid, int32_t error);
 void PID_Reset(_sPID *pid);
